[refactor]: dto 위치 변환 작업
