+++ src/main/java/com/nodaji/payment/dto/request/WithdrawRequestDto.java
@@ -1,5 +1,9 @@
 package com.nodaji.payment.dto.request;
 
+import com.nodaji.payment.global.domain.entity.History;
+
+import java.util.Date;
+
 public record WithdrawRequestDto (
         Long price,
         Long charge,
@@ -7,4 +11,12 @@ public record WithdrawRequestDto (
         String accountNum,
         String accountOwnerName
 ){
+    public History toEntity(String userId, WithdrawRequestDto req){
+        return History.builder()
+                .transaction_status("출금")
+                .price(req.price()+req.charge())
+                .createdAt(new Date())
+                .userId(userId)
+                .build();
+    }
 }
